HV: Modularize iommu
Modified some of the code in order to isolate
iommu code to vtd.c.
Treating acrn_vm as higher level module, removed
any dependency of acrn_vm structure from vtd.c.
Removed any reference to snoopy_mem from iommu_structure,
and added code to set and read iommu_snoop, which is
limited inside the iommu structure.

Tracked-On: #1842
