+++ hypervisor/boot/dmar_parse.c
@@ -7,7 +7,7 @@
 #ifdef CONFIG_DMAR_PARSE_ENABLED
 #include <hypervisor.h>
 #include "pci.h"
-#include "vtd.h"
+#include <vtd.h>
 #include "acpi_priv.h"
 
 enum acpi_dmar_type {
