+++ hypervisor/dm/vpci/pci_pt.c
@@ -31,6 +31,7 @@
 
 #include <hypervisor.h>
 #include "pci_priv.h"
+#include <vtd.h>
 
 static inline uint32_t pci_bar_base(uint32_t bar)
 {
