+++ hypervisor/include/arch/x86/guest/vm.h
@@ -162,7 +162,6 @@ struct acrn_vm {
 	spinlock_t softirq_dev_lock;
 	struct list_head softirq_dev_entry_list;
 	uint64_t intr_inject_delay_delta; /* delay of intr injection */
-	bool snoopy_mem;
 } __aligned(PAGE_SIZE);
 
 #ifdef CONFIG_PARTITION_MODE
