+++ hypervisor/include/arch/x86/hv_arch.h
@@ -36,7 +36,6 @@
 #include <vmx.h>
 #include <vmcs.h>
 #include <assign.h>
-#include <vtd.h>
 
 #include <guest.h>
 #include <vmexit.h>
